Refactored regex::Regex type into its own module.
The new module is the handle module which reflects Regex's role as the
handle to RegexKind's body (in the handle/body idiom).
